fix: rename standalone to headless
