+++ src/errors/HeadlessInitializeError.ts
