+++ src/errors/index.ts
@@ -1,3 +1,3 @@
 export { default as FetchError } from "./FetchError";
 export { default as HeadlessExitedError } from "./HeadlessExitedError";
-export { default as HeadlessInitializeError } from "./StandaloneInitializeError";
+export { default as HeadlessInitializeError } from "./HeadlessInitializeError";
